+++ biopet-utils/src/main/scala/nl/lumc/sasc/biopet/utils/package.scala
@@ -100,4 +100,17 @@ package object utils {
       case _ => None
     }
   }
+
+  /** Function to sort Any values */
+  def sortAnyAny(a: Any, b: Any): Boolean = {
+    a match {
+      case ai: Int =>
+        b match {
+          case bi: Int    => ai < bi
+          case bi: Double => ai < bi
+          case _          => a.toString < b.toString
+        }
+      case _ => a.toString < b.toString
+    }
+  }
 }
